Script robustification and message bus integration (#119)
Script robustification and message bus integration
=====================================

Script robustification
---------------------

The installation and session running scripts have been greatly improved and made more robust - reduced redundancy, better error handling, more modularity using functions, style improvements, bugfixes. The indent width has been increased from 2 to 4 in all top-level scripts.

Message bus integration
-------------------------

The discrete events implementation has been refactored to publish the events as messages to an MQTT message broker instead of outputting the events to file directly. A new package, `edu.arizona.tomcat.Events` has been implemented, and collects the classes related to event handling (except for the iron door opening event, which is in the ASISTBlocks package since it fits in better over there because of the block registration requirement.

The handler class for events that are fired on the Forge EVENT_BUS bus is `ForgeEventHandler`. To handle a new type of Forge event, add a new polymorphic `handle` method with the argument as the appropriate Forge event. To change the behavior of an already handled Forge event, or to split the functionality into additional branches based on the type of the objects involved in the event (for some `BlockInteraction` events we may want to publish the event to a separate topic, adding additional information about the event, e.g. for door opening in the USAR missions).

The current topic hierarchy is now
- `observations`
  - `state`
  - `events`
    - `mob_attacked`
    - `block_interaction`
    - `entity_death`
    - `iron_door_opened`

### Message examples

Example of a `block_interaction` event:

```json
{
    "player_name": "tomcat",
    "block_position": {
        "x": 56.0,
        "y": 65.0,
        "z": 65.0
    },
    "block_type": "net.minecraft.block.BlockButtonStone",
    "timestamp": "2020-04-23T15:45:23:558Z",
    "event_type": "block_interaction"
}
```


Example of a `mob_attacked` event:

```json
{
    "target_position": {
        "x": 62.095616864133056,
        "y": 64.29878348147426,
        "z": 60.40173640449803
    },
    "target_velocity": {
        "x": 0.138,
        "y": 0.275125,
        "z": -0.117125
    },
    "player_name": "tomcat",
    "target_type": "net.minecraft.entity.monster.EntityZombie",
    "target_id": "b01ec5a8-bcab-4a35-910c-259997d392c5",
    "target_health": 18.119998931884766,
    "timestamp": "2020-04-23T17:32:51:997Z",
    "event_type": "mob_attacked"
}
```

Note: In the above schema, we are retaining the full path to the block class, so that we can clearly distinguish between built-in Minecraft blocks and custom block classes, e.g. `edu.arizona.tomcat.ASISTBlocks.BlockAsistIron`. We can change this later if desired.

Examples of an `entity_death` event 

For a zombie:

```json
{
    "entity_type": "net.minecraft.entity.monster.EntityZombie",
    "entity_name": "Zombie",
    "entity_id": "36af5aea-ccbc-4332-9933-c21d49b390fa",
    "timestamp": "2020-04-23T17:36:31:980Z",
    "event_type": "entity_death"
}
```

For a player:

```
{
    "entity_type": "net.minecraft.entity.player.EntityPlayerMP",
    "entity_name": "tomcat",
    "entity_id": "578193db-da09-3d22-8eb9-a1d3ebc7e3f3",
    "timestamp": "2020-04-23T17:50:37:538Z",
    "event_type": "entity_death"
}
```


If we need to publish messages to the message bus from somewhere outside the `ForgeEventHandler` class, we can just get the singleton instance of the `MqttService` class and publish the event using it. There is an overloaded `publish` method of that class that can automatically publish JSON serializations of Java objects that can be serialized using Gson.

Other changes
---------------

### Major

Tutorial data is now being recorded as well in addition to the main mission data.
The option to disable MQTT has been removed and the scripts now launch the `mosquitto` message broker on the default port if it is not already running.

### Minor

- A/V files (.mpg and .wav) files have been added to `.gitignore`.
- Removed unneccessary comments in `build.gradle`.
- Spruced up comments including typo fixes.
- Removed some unnecessary imports in some `.java` files.
- Changed instances of 'he/them' to the gender-neutral pronouns 'they/them'. 
- Factored out timestamping into a separate utility class, `TimeStamper`.
- Increased time precision of timestamps to milliseconds.
- Spruced up instructions in the tutorial mission.
- Removed `CommandEvents` event handler since it was making the log files really huge.